+++ packages/react/src/composite/Pagination/src/hooks/usePagination/index.ts
@@ -1,0 +1,2 @@
+export { usePagination } from "./usePagination";
+export { DOTS } from "./usePagination.definitions";