+++ packages/react/src/composite/Pagination/src/hooks/usePagination/usePagination.definitions.ts
@@ -1,0 +1,10 @@
+export const DOTS = "...";
+
+export const range = (start: number, end: number) => {
+  const length = end - start + 1;
+  /*
+  	Create an array of certain length and set the elements within it from
+    start value to end value.
+  */
+  return Array.from({ length }, (_, idx) => idx + start);
+};