+++ packages/react/src/composite/Pagination/src/hooks/usePagination/usePagination.types.ts
@@ -1,0 +1,8 @@
+import { PaginationProps } from "../../pagination.types";
+
+export type UsePaginationProps = Pick<
+  PaginationProps,
+  "activePage" | "pageCount"
+> & {
+  siblingCount?: number;
+};