+++ packages/react/src/composite/Pagination/src/index.ts
@@ -1,0 +1,6 @@
+import { Pagination } from "./Pagination";
+import "@nimbus-ds/styles/packages/composite/pagination/index.css";
+
+export { Pagination } from "./Pagination";
+export type { PaginationProps } from "./pagination.types";
+export default Pagination;