+++ packages/react/src/composite/Pagination/src/pagination.definitions.ts
@@ -1,0 +1,2 @@
+export const generateKey = (pageNumber: number | string, index: number) =>
+  `${pageNumber}-${index}`;