+++ packages/react/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This package is intended for internal use in generating builds of each design system package. It contains all the necessary settings and dependencies to optimize the creation of our builds.
 
+## 2022-12-20 `1.1.0`
+
+### 🎉 New features
+
+- Added new composite `Pagination` component. ([#68](https://github.com/TiendaNube/nimbus-design-system/pull/68) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2022-12-15 `1.0.2`
 
 ### 🐛 Bug fixes